pyproject.toml
[tool.poetry]
name = "oblivion"
packages = [{include = "workflow"}]
version = "0.1.0"
description = "Project Oblivion: Inducing Controlled Hallucinations in AI Models"
authors = ["r1char9"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
numpy = "^2.2.4"
transformers = "^4.49.0"
datasets = "^3.4.1"
hydra-core = "^1.3.2"
openai = "^1.66.3"
deepeval = "^2.5.9"
langid = "^1.1.6"
langgraph = "^0.3.21"
langchain-ollama = "^0.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
